requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
langgraph
requests
pydantic
python-multipart
mlflow
python-dotenv
aiofiles
tqdm
PyPDF2
langchain 
langchain-community
streamlit
langchain-classic
langchain_huggingface
unstructured
